requirements.txt
pandas
matplotlib
streamlit
tensorflow
Pillow
tf-keras
torch
huggingface_hub